fix test fixture breakages due to size estimation
In commit ad11546 we broke an old SizeEstimation
assumption - that an address is used for size estimation,
which should be treated as a 25 byte script.

The address is now handled as a script, so estimations
are more correct, but this caused some test fixtures to
break.

The two cases involved size esimations in which a change
output would be dropped if adding the output cost more in fee,
and the case where an economical change output is created
although it costs 0.0001. Both of these relied on the size
estimation placing them within $outputNumberOfBytes to the
1000byte mark for base fee.